+++ internal/migrate/migrate,go.go
@@ -8,7 +8,7 @@ import (
 
 func RunMigrations(dbURL, migrationsPath string) error {
 	m, err := migrate.New(
-		"file:///Users/konoko/Documents/Go/Homework3/study/migrations"+migrationsPath,
+		"file://"+migrationsPath,
 		dbURL,
 	)
 	if err != nil {
@@ -17,13 +17,12 @@ func RunMigrations(dbURL, migrationsPath string) error {
 		//уточнить
 	}
 	err = m.Up()
-	if err != nil {
-		if errors.Is(err, migrate.ErrNoChange) {
-			log.Printf("Nothing to migrate")
-			return nil
-		}
-		log.Printf("Failed to migrate: %v", err)
+
+	if errors.Is(err, migrate.ErrNoChange) {
+		log.Printf("Nothing to migrate")
+		return nil
 	}
+
 	log.Printf("Migrations successfully migrated")
 	return nil
 }
