--- migrate/000001_create_users_table.down.sql
@@ -1,3 +1,0 @@
-CREATE TABLE IF NOT EXISTS users (
-     id TEXT PRIMARY KEY
-);
