--- migrate/000001_create_users_table.up.sql
@@ -1,1 +1,0 @@
-DROP TABLE IF EXISTS users;
